feat: GET /user/teams API 추가
- /user API 코드 구조 리팩토링
- GET /user/teams API 추가
